fixes and plugins changes
[*]: Added event="VeryLazy" to speed up booting
[ui_extra]: temp fold fixes
[filemanager]: added oil plugin
[coding]: moved blink, treesitter, and commentary
[navigation]: moved tmux navigator, fzf, and neotree
[ui]: moved nui
[formatter]: added conform plugin to replace none-ls
[non-lsp]: moved bats, and rainbow_csv
