+++ lua/config/ui_extra.lua
@@ -1,0 +1,10 @@
+local function setup_other()
+	vim.o.foldcolumn = "1"
+	vim.o.foldlevel = 99
+	vim.o.foldlevelstart = 99
+	vim.o.foldenable = true
+end
+
+return {
+	setup = setup_other,
+}
