--- lua/plugins/core/commentary.lua
@@ -1,3 +1,0 @@
-return {
-	"tpope/vim-commentary",
-}
