--- lua/plugins/core/fuzzy.lua
@@ -1,12 +1,0 @@
-return {
-	"ibhagwan/fzf-lua",
-	dependencies = { "nvim-tree/nvim-web-devicons" },
-	keys = {
-		{ "<leader>ff", "<cmd>FzfLua files<cr>", desc = "Find Files" },
-		{ "<leader>fr", "<cmd>FzfLua resume<cr>", desc = "Resumes previous operation" },
-		{ "<leader>fg", "<cmd>FzfLua live_grep<cr>", desc = "Live Grep" },
-		{ "<leader>fb", "<cmd>FzfLua buffers<cr>", desc = "Buffers" },
-		{ "<leader>fh", "<cmd>FzfLua helptags<cr>", desc = "Show help tags" },
-		{ "<leader>fc", ":FzfLua ", desc = "Start FzfLua command line" },
-	},
-}
