+++ lua/plugins/core/git.lua
@@ -1,9 +1,10 @@
 return {
-	{ "tpope/vim-fugitive" },
-	{ "sindrets/diffview.nvim" },
+	{ "tpope/vim-fugitive", event = "VeryLazy" },
+	{ "sindrets/diffview.nvim", event = "VeryLazy" },
 	{
 		"lewis6991/gitsigns.nvim",
 		dependencies = { "nvim-lua/plenary.nvim" },
+		event = "VeryLazy",
 		opts = {
 			numhl = true,
 			current_line_blame_opts = {
