--- lua/plugins/core/nui.lua
@@ -1,4 +1,0 @@
-return {
-	"MunifTanjim/nui.nvim",
-	lazy = true,
-}
