--- lua/plugins/core/tmux.lua
@@ -1,3 +1,0 @@
-return {
-	"christoomey/vim-tmux-navigator",
-}
