+++ lua/plugins/extra/limelight.lua
@@ -1,5 +1,6 @@
 return {
 	"junegunn/limelight.vim",
+	event = "VeryLazy",
 	keys = {
 		{ "<leader>ll", "<cmd>Limelight!!<cr>", desc = "Toggle limelight" },
 	},
