+++ lua/plugins/lsp/non-lsp.lua
@@ -1,0 +1,23 @@
+return {
+	{
+		"cameron-wags/rainbow_csv.nvim",
+		event = "VeryLazy",
+		config = true,
+		ft = {
+			"csv",
+			"tsv",
+			"csv_semicolon",
+			"csv_whitespace",
+			"csv_pipe",
+			"rfc_csv",
+			"rfc_semicolon",
+		},
+		cmd = {
+			"RainbowDelim",
+			"RainbowDelimSimple",
+			"RainbowDelimQuoted",
+			"RainbowMultiDelim",
+		},
+	},
+	{ "aliou/bats.vim", event = "VeryLazy" },
+}
