--- lua/plugins/nonlsp/bats.lua
@@ -1,3 +1,0 @@
-return {
-	"aliou/bats.vim",
-}
