--- lua/plugins/nonlsp/csv.lua
@@ -1,19 +1,0 @@
-return {
-	"cameron-wags/rainbow_csv.nvim",
-	config = true,
-	ft = {
-		"csv",
-		"tsv",
-		"csv_semicolon",
-		"csv_whitespace",
-		"csv_pipe",
-		"rfc_csv",
-		"rfc_semicolon",
-	},
-	cmd = {
-		"RainbowDelim",
-		"RainbowDelimSimple",
-		"RainbowDelimQuoted",
-		"RainbowMultiDelim",
-	},
-}
