+++ lua/plugins/plugins.lua
@@ -3,6 +3,5 @@ return {
 	{ import = "plugins.exp" },
 	{ import = "plugins.extra" },
 	{ import = "plugins.lsp" },
-	{ import = "plugins.nonlsp" },
 	{ import = "plugins.util" },
 }
