+++ lua/plugins/util/debugger.lua
@@ -8,6 +8,7 @@ return {
 		"nvim-neotest/nvim-nio",
 		"williamboman/mason.nvim",
 	},
+	event = "VeryLazy",
 	config = function()
 		local dap, ui = require("dap"), require("dapui")
 		local before = dap.listeners.before
