+++ lua/plugins/util/surround.lua
@@ -1,10 +1,12 @@
 return {
 	{
 		"tpope/vim-surround",
+		event = "VeryLazy",
 		dependencies = { "tpope/vim-repeat" },
 	},
 	{
 		"echasnovski/mini.ai",
+		event = "VeryLazy",
 		branch = "stable",
 		config = function()
 			require("mini.ai").setup({
